% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptimalTSPTour.R
\name{getOptimalTSPTour}
\alias{getOptimalTSPTour}
\title{Get optimal TSP tour from file.}
\usage{
getOptimalTSPTour(prob.name, path.to.tsplib)
}
\arguments{
\item{prob.name}{[\code{string}]\cr
Problem name.}

\item{path.to.tsplib}{[\code{filepath}]\cr
Absolute/relative path to tsplib instances and their tour files.}
}
\value{
[\code{integer}] Permutation.
}
\description{
Given a problem name, e.g. \dQuote{a280} and a path to a
directory with \dQuote{.opt.tour} files in tsplib tour format the function
reads the file and returns the permutation.
}
