% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTourByIncreasingWeight.R
\name{getTourByIncreasingWeight}
\alias{getTourByIncreasingWeight}
\title{Tour where nodes are visited by increasing weight.}
\usage{
getTourByIncreasingWeight(prob, packing)
}
\arguments{
\item{prob}{[\code{ttp_instance}]\cr
TTP instance.}

\item{packing}{[\code{integer}]\cr
Packing vector.}
}
\value{
[\code{integer}] Permutation of nodes.
}
\description{
For the node-weight dependent TSP (WTSP) in particular and
for the more realistic multi-component traveling thief problem (TTP) it
is reasonable to think of heavy-weights nodes being visited late in the
optimal tour. This function calculates the extreme setting. I.e. starting
in node 1 the remaining nodes are visited in increasing order of their
node weight, i.e. the sum of weights of packed items.
}
