% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{TTP instance loader.}
\usage{
load(path)
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path to TTP instance.}
}
\value{
[\code{ttp_instance}] Basically a list with class \dQuote{ttp_instance}.
}
\description{
Loads TTP instance.
}
