% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPacking.R
\name{loadPackings}
\alias{loadPackings}
\title{Load packing(s)}
\usage{
loadPackings(path, which = NULL)
}
\arguments{
\item{path}{[\code{string}]\cr
File path to packings.}

\item{which}{[\code{integer}]\cr
Which packings should be returned?
Default is \code{NULL} which means all packings.}
}
\description{
Given a path to a file with packings, i.e. rowwise strings
from \eqn{\{0, 1\}} the function imports the packings and returns a matrix.
}
