% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPacking.R
\name{makeRandomPacking}
\alias{makeRandomPacking}
\title{Get random packing plan}
\usage{
makeRandomPacking(prob, p)
}
\arguments{
\item{prob}{[\code{ttp_instance}]\cr
TTP instance.}

\item{p}{[\code{numeric}]\cr
With probability \eqn{p \in (0, 1]} each single item is packed.}
}
\value{
[\code{integer}] Vector in \eqn{\{0,1\}^m} where \eqn{m}
is equal to \code{prob$m}.
}
\description{
Get random packing plan
}
