% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector.}
\usage{
normalize(x)
}
\arguments{
\item{x}{[\code{numeric}]\cr
Numeric input vector.}
}
\value{
[\code{numeric}] Normalized \code{x}.
}
\description{
After normalization, all values in \code{normalize(x)} are
in \eqn{[0, 1]}.
}
