% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{objectives}
\alias{objectives}
\alias{tsp}
\alias{wtsp}
\alias{ttp}
\title{Objectives}
\usage{
tsp(tour, prob, packing, ...)

wtsp(tour, prob, packing, ...)

ttp(tour, prob, packing, ...)
}
\arguments{
\item{tour}{[\code{integer}]\cr
Permutation of nodes.}

\item{prob}{[\code{ttp_instance}]\cr
TTP instance.}

\item{packing}{[\code{integer}]\cr
Binary vector, i.e. with ones and zeros, of length \code{prob$m}
where each component indicates whether the corresponding item
should be packed (1) or not (0).
This argument is ignored by \code{tsp}.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{numeric(1)}] Objective function value.
}
\description{
Classical Traveling Salesperson Problem (tsp),
node-weighted TSP (wtsp) and single-objective formulation of
the Traveling-Thief-Problem (ttp).
}
