% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWTSPSolver.R
\name{runWTSPSolver}
\alias{runWTSPSolver}
\title{Run weighted TSP solver.}
\usage{
runWTSPSolver(instance, packing, tours = NULL, mu = 1L,
  mutation = "swap", survival.strategy = "classic", max.evals)
}
\arguments{
\item{instance}{[\code{character(1)}]\cr
Path to instance.}

\item{packing}{[\code{integer}]\cr
Binary vector of packed items.}

\item{tours}{[\code{matrix}]\cr
A \code{mu} times \eqn{n} matrix where \code{mu} is the population size and
\code{n} is the number of nodes. Each row is one initial TSP tour.
If \code{NULL}, random permutation are generated.}

\item{mu}{[\code{integer(1)}]\cr
Population size.}

\item{mutation}{[\code{character(1)}]\cr
One of \dQuote{swap}, \dQuote{jump}, \dQuote{inversion} or \dQuote{scramble}.}

\item{survival.strategy}{[\code{character(1)}]\cr
How to perform survival selection. Option \dQuote{classic} drops the
worst out of \eqn{\mu + 1} individuals while option \dQuote{parent} compares
the fitness of the child with its parent fitness only.
Note that for \code{mu = 1} both strategies behave in the same way.
Default is \code{classic}.}

\item{max.evals}{[\code{integer}]\cr
Maximum number of function evaluations.}
}
\value{
[\code{list}]
}
\description{
Given a problem instance and a binary packing plan, i.e. with \code{packing[i] == 1}
meaning that the i-th element must be taken, calculate a minimum length node-weighted
TSP tour.
}
