% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWTSPSolverInDynamicSetting.R
\name{runWTSPSolverInDynamicSetting}
\alias{runWTSPSolverInDynamicSetting}
\title{Wrapper for runWTSPSolver taking care of dynamic changes.}
\usage{
runWTSPSolverInDynamicSetting(pathToInstance, pathToPackings, mutation, mu,
  survival.strategy, tau, max.evals.initial, max.evals = NULL,
  n.changes = 3L, start.from.scratch = FALSE, full = FALSE)
}
\arguments{
\item{pathToInstance}{[\code{character(1)}]\cr
Path to instance.}

\item{pathToPackings}{[\code{character(1)}]\cr
Path to file with row-wise binary packing vector.}

\item{mutation}{[\code{character(1)}]\cr
See \code{\link{runWTSPSolver}}.}

\item{mu}{[\code{integer(1)}]\cr
Population size.}

\item{survival.strategy}{[\code{character(1)}]\cr
How to perform survival selection. Option \dQuote{classic} drops the
worst out of \eqn{\mu + 1} individuals while option \dQuote{parent} compares
the fitness of the child with its parent fitness only.
Note that for \code{mu = 1} both strategies behave in the same way.
Default is \code{classic}.}

\item{tau}{[\code{integer(1)}]\cr
Dynamic change in packing is triggered every \code{tau} function evaluations.}

\item{max.evals.initial}{[\code{integer(1)}]\cr
Number of evaluations for first initial run.}

\item{max.evals}{[\code{integer(1)} | \code{NULL}]\cr
Number of evaluations used after each dynamic change. Default is \code{NULL}.
In this case we have \code{max.evals = tau}.}

\item{n.changes}{[\code{integer(1)}]\cr
Number of dynamic changes.}

\item{start.from.scratch}{[\code{logical(1)}]\cr
Start with random population each time a dynamic change occurs?
Default is \code{FALSE}.}

\item{full}{[\code{logical(1)}]\cr
If \code{TRUE} a list of length \code{n.changes + 1} is returned with components
\dQuote{tour}, \dQuote{tour.length}, \dQuote{toursFinal} and \dQuote{trajectory}.
If \code{FALSE} - the default - only a length-\code{n.changes + 1} numeric vector
of tour length is returned.}
}
\value{
[\code{list} | \code{numeric}] See \code{full} for details
}
\description{
Wrapper for runWTSPSolver taking care of dynamic changes.
}
