% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftTour.R
\name{shiftTour}
\alias{shiftTour}
\title{Shift tour.}
\usage{
shiftTour(x)
}
\arguments{
\item{x}{[\code{integer}]\cr
Permutation of \eqn{\{1, \ldots, n\}}.}
}
\value{
[\code{integer}] Shifted permutation.
}
\description{
Shifting is performed such that node 1 is the first in the
permutation.
}
