% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{writeProblem}
\alias{writeProblem}
\title{TTP instance writer.}
\usage{
writeProblem(x, path, overwrite = FALSE, ...)
}
\arguments{
\item{x}{[\code{ttp_instance}]\cr
TTP instance.}

\item{path}{[\code{character(1)}]\cr
Path to file.}

\item{overwrite}{[\code{logical(1)}]\cr
Should file be overwritten if it already exists?
Default is \code{FALSE}.}
}
\value{
[\code{invisible(TRUE)}]
}
\description{
Writes TTP instance to file.
}
